+++ changelog.md
@@ -2,10 +2,12 @@
 - Add Creative Wrench
   - Links any 2 entities together for warping
   - Links On/Off Switches to toggleable blocks
+  - Opens ? Block, Invisible ? Block, or Storage Bricks GUI
+- Add new GUI to set a refill timer for ? Blocks, Invisible ? Blocks, or Storage Bricks
+- Add new wrench sounds
 - Add Immersive Paintings support for warping
 - Add Twilight Forest magic painting support for warping
 - Add chipped mushrooms to bouncy tag
-- Add new wrench sounds
 - Throw shells faster depending on movement, such as from vehicles
 - Pipes now will keep the player on the vehicle when warping
 - When a pipe is sideways, you will teleport facing away from the destination pipe